+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## Unreleased
+
+### Added
+
+- Added missing `jsonapi:authorizer` generator command.
+
 ## [1.0.0-alpha.3] - 2021-02-09
 
 ### Added